Changed Fread variables to long.

+++ include/dlb.h
@@ -72,7 +72,7 @@ void NDECL(dlb_cleanup);
 
 dlb *FDECL(dlb_fopen, (const char *, const char *));
 int FDECL(dlb_fclose, (DLB_P));
-int FDECL(dlb_fread, (char *, long, long, DLB_P));
+long FDECL(dlb_fread, (char *, long, long, DLB_P));
 int FDECL(dlb_fseek, (DLB_P, long, int));
 char *FDECL(dlb_fgets, (char *, int, DLB_P));
 int FDECL(dlb_fgetc, (DLB_P));
@@ -84,7 +84,7 @@ boolean rsrc_dlb_init(void);
 void rsrc_dlb_cleanup(void);
 boolean rsrc_dlb_fopen(dlb *dp, const char *name, const char *mode);
 int rsrc_dlb_fclose(dlb *dp);
-int rsrc_dlb_fread(char *buf, size_t size, size_t quan, dlb *dp);
+long rsrc_dlb_fread(char *buf, long size, long quan, dlb *dp);
 int rsrc_dlb_fseek(dlb *dp, long pos, int whence);
 char *rsrc_dlb_fgets(char *buf, int len, dlb *dp);
 int rsrc_dlb_fgetc(dlb *dp);
